+++ src/app/auth/components/login-form/login-form.component.ts
@@ -34,7 +34,7 @@ export class LoginFormComponent implements OnInit {
     });
   }
 
-  ngOnInit(): void { }
+  ngOnInit(): void {}
 
   submit() {
     // Inputed data