+++ src/environments/environment.dev.ts
@@ -2,5 +2,6 @@ export const environment = {
   production: true,
   baseUrl: `https://evilflowers-dev.fiit.stuba.sk`,
   assetsUrl: `https://elvira-dev.fiit.stuba.sk/assets`,
-  catalog_id: `b349b5a1-4d50-4bed-bf8d-c56d6c7b1280`
+  catalog_id: `b349b5a1-4d50-4bed-bf8d-c56d6c7b1280`,
+  httpRequestInterval: 300000, // 300s
 };